requirements.txt
streamlit
tensorflow-cpu
Pillow
numpy
matplotlib
scikit-learn
opencv-python
python-dotenv